+++ poincare/src/layout/layout_span_decoder.h
@@ -21,17 +21,17 @@ class LayoutSpanDecoder : public ForwardUnicodeDecoder {
 
   LayoutSpanDecoder(const Rack* rack, size_t initialPosition = 0,
                     size_t lastPosition = k_noSize)
-      : LayoutSpanDecoder(initialPosition == 0
-                              ? static_cast<const Layout*>(rack->nextNode())
-                          /* initialPosition is allowed to be lastPosition for
-                           * convenience, in that case m_layout is nullptr and
-                           * should not be dereferenced */
-                          : initialPosition < rack->numberOfChildren()
-                              ? rack->child(initialPosition)
-                              : nullptr,
-                          (lastPosition == k_noSize ? rack->numberOfChildren()
-                                                    : lastPosition) -
-                              initialPosition) {}
+      : LayoutSpanDecoder(
+            initialPosition == 0 ? static_cast<const Layout*>(rack->nextNode())
+            /* initialPosition is allowed to be lastPosition for
+             * convenience, in that case m_layout is nullptr and
+             * should not be dereferenced */
+            : static_cast<int>(initialPosition) < rack->numberOfChildren()
+                ? rack->child(initialPosition)
+                : nullptr,
+            (lastPosition == k_noSize ? rack->numberOfChildren()
+                                      : lastPosition) -
+                initialPosition) {}
 
   LayoutSpanDecoder(LayoutSpan span)
       : LayoutSpanDecoder(span.start, span.length) {}
